Using cookies
Used cookies to maintain login and logout state and added mailID to uploaded files which gives list of all files uploaded by a user and in this we can display all files uploaded by a particular user. To do next: when user login/logout and if user is on .../myFiles and the component don't get refreshed i.e. component doesn't get re-rendered on changin login and logout state.